requirements.txt
absl-py==2.1.0
aiohttp==3.8.4
aiosignal==1.3.1
alabaster==0.7.16
albumentations==1.4.2
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
blinker==1.7.0
boto3==1.34.66
botocore==1.34.66
build==1.1.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
contourpy==1.2.0
coverage==5.3.1
cycler==0.12.1
Deprecated==1.2.14
dnspython==2.6.1
docutils==0.17.1
einops==0.3.2
filelock==3.6.0
fire==0.6.0
flatbuffers==24.3.7
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.3.1
future==1.0.0
grpcio==1.62.1
h5py==3.10.0
humanfriendly==10.0
hydra-core==1.3.2
idna==3.6
ifaddr==0.2.0
imageio==2.34.0
imagesize==1.4.1
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
json-tricks==3.16.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.3
lightning-utilities==0.3.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
numpy==1.26.4
nvidia-ml-py3==7.352.0
omegaconf==2.3.0
onnx==1.13.0
onnx-simplifier==0.4.36
onnxruntime==1.13.1
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==24.0
pandas==2.2.1
pillow==10.2.0
pip-tools==7.4.1
prettytable==3.10.0
protobuf==3.20.3
psutil==5.8.0
py-eureka-client==0.9.5
pycocotools==2.0.6
pyctuator==0.13.2
pyDeprecate==0.3.2
Pygments==2.17.2
pyparsing==2.4.5
pyproject_hooks==1.0.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.4
pytorch-lightning==1.8.2
pytorch-quantization==2.2.1
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.6.2
referencing==0.34.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
s3transfer==0.10.1
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
snowballstemmer==2.2.0
Sphinx==4.0.3
sphinx-rtd-theme==1.3.0
sphinx_glpi_theme==0.6
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
stringcase==1.2.0
super-gradients==3.1.3
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==1.1.0
threadpoolctl==3.3.0
tifffile==2024.2.12
tomli==2.0.1
torch==1.13.1+cu117
torch-tb-profiler==0.4.3
torchaudio==0.13.1+cu117
torchmetrics==0.8.0
torchvision==0.14.1+cu117
tqdm==4.66.2
treelib==1.6.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==2.2.2
wget==3.0
wrapt==1.16.0
yarl==1.8.2